WIP: Generator - Call ops successfully. Uses hard-coded values.
Program Counter and return values in the switch statement is hard-coded. Additionally, the number of cycles is not properly accumulated. Initial tests show that execution functions properly for the first piece of the BIOS, with Program Counter and Stack Pointer incrementing after single call to execute(sys_t*).

The Program Counter does not seem to wind up at the expected place, coming up as 0x06 when it should likely be 0x07 or 0x0A. This may signify an error in the JRNZn instruction at 0x0A, or that execution is not terminating at a landing/terminator.